+++ download_premarket.py
@@ -9,36 +9,57 @@ today = datetime.now().strftime("%Y-%m-%d")
 # script directory ... Project/Python
 base_path = Path(__file__).resolve().parent
 
-# ETF.txt file path
-tickers_file_path = base_path.parent / "Batch" / "Premarket.txt"
-
-# ingest tickers from file
-with open(tickers_file_path, "r") as file:
-    tickers = [line.strip() for line in file if line.strip()]
-
 # download tickers to data folder
 data_folder = base_path.parent.parent / "data" / "finance" / "intraday"
-for ticker in tickers:
-    data = yf.download(ticker,period="1d",interval="5m",prepost=True,progress=False)
+data_folder.mkdir(parents=True, exist_ok=True)
 
-    data = data.tz_convert("America/New_York")
-    data.index = data.index.tz_localize(None)
+# ticker file paths
+equity_file_path = base_path.parent / "Batch" / "Equity.txt"
+etf_file_path = base_path.parent / "Batch" / "ETF.txt"
 
-    if hasattr(data.columns, "levels"):
-        data.columns = data.columns.get_level_values(0)
+for tickers_file_path in [equity_file_path, etf_file_path]:
 
-    # keep only premarket
-    premarket = data.between_time("04:00", "09:30")
+    if not tickers_file_path.exists():
+        print(f"Missing file: {tickers_file_path}")
+        continue
 
-    # so that indices like ^DJI and ^GSPC can be included in Intraday.txt
-    if ticker.startswith("^"):
-        filename = ticker[1:] + "-i.csv"
-        print(f"{ticker[1:]}")
-    else:
-        filename = ticker + "-i.csv"
-        print(f"{ticker}")
-    filepath = data_folder / filename
-    premarket.to_csv(filepath)
+    # ingest tickers from file
+    with open(tickers_file_path, "r") as file:
+        tickers = [line.strip() for line in file if line.strip()]
 
-    # give server more time to process requests
-    time.sleep(2)
+    for ticker in tickers:
+
+        try:
+
+            data = yf.download(
+                ticker,
+                period="1d",
+                interval="5m",
+                prepost=True,
+                progress=False
+            )
+
+            data = data.tz_convert("America/New_York")
+            data.index = data.index.tz_localize(None)
+
+            if hasattr(data.columns, "levels"):
+                data.columns = data.columns.get_level_values(0)
+
+            # keep only premarket
+            premarket = data.between_time("04:00", "09:30")
+
+            # so that indices like ^DJI and ^GSPC can be included in Intraday.txt
+            if ticker.startswith("^"):
+                filename = ticker[1:] + "-i.csv"
+                print(f"{ticker[1:]}")
+            else:
+                filename = ticker + "-i.csv"
+                print(f"{ticker}")
+            filepath = data_folder / filename
+            premarket.to_csv(filepath)
+
+        except Exception as e:
+            print(f"{ticker}: {e}")
+
+        # give server more time to process requests
+        time.sleep(2)
